Replace todo project with workout project
I'm planning to work through the material in Elixir in Action, but I
figured I'd try to make it a little more real and relevant by extending
the demo for my own interests. I'm already pretty happy with my todo app
(Things), so I figured if I can translate the demo application into a
workout app then I can use it as a base to build off of.